apps/fast_f1_service/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "fast_f1_service" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = """
 --cov
 --cov-report html:'../../coverage/apps/fast_f1_service/html'
 --cov-report xml:'../../coverage/apps/fast_f1_service/coverage.xml'
 --html='../../reports/apps/fast_f1_service/unittests/html/index.html'
 --junitxml='../../reports/apps/fast_f1_service/unittests/junit.xml'
"""

[tool.hatch.build.targets.wheel]
packages = [ "fast_f1_service" ]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
exclude = [ ".ruff_cache", ".svn", ".tox", ".venv", "dist" ]
line-length = 88
indent-width = 4

  [tool.ruff.lint]
  select = [ "E", "F", "UP", "B", "SIM", "I" ]
  ignore = [ ]
  fixable = [ "ALL" ]
  unfixable = [ ]

[project]
name = "fast_f1_service"
version = "1.0.0"
description = "Automatically generated by Nx."
requires-python = ">=3.9,<4"
readme = "README.md"
dependencies = [
 "fastf1>=3.5.3",
 "grpcio>=1.71.0",
 "grpcio-tools>=1.71.0",
]

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"
